+++ CatBot/help/__init__.py
@@ -2,4 +2,4 @@
 Commands for the /help command.
 """
 
-from .help import PRIVATE, PUBLIC
+from .commands import PRIVATE, PUBLIC